### 2024-04-22 #176 @jinningwang
Submission for ANDES
Submitting Author: Jinning Wang (@jinningwang)
All current maintainers: Hantao Cui (@cuihantao), Jinning Wang (@Jinning Wang)
Package Name: ANDES
One-Line Description of Package: Power system transient dynamics simulation with symbolic modeling and numerical analysis
Repository Link:  https://github.com/CURENT/andes
Version submitted: v1.9.1
Version accepted: TBD

---

ANDES is an open-source Python library for power system modeling, computation, analysis, and control, serving as the core simulation engine for the CURENT Large scale Testbed (LTB). It supports power flow calculation, transient stability simulation, and small-signal stability analysis for transmission systems. ANDES implements a symbolic-numeric framework for rapid prototyping of differential-algebraic equation-based models.

> [^1]: Please fill out a pre-submission inquiry before submitting a data visualization package.

Power system researchers and engineers.
Power system transient stability modeling and analysis.

To the best of my knowledge, ANDES is the only python package, for power system transient stability modeling and simulation, with the proposed hybrid symbolic-numeric method.

The
 [editor template can be found here][Editor Template].
