### 2024-04-26 #176 @Batalex
Hey,
Thank you for your trust in our process by submitting your second package to pyOpenSci.
I want to put this submission on hold for now while the first one is ongoing. This will enable us to manage most effectively our volunteers' time, and it could be a could opportunity to keep the first team around for this submission.

Please let me know if this arrangement works for you.
