### 2024-05-26 c2f9a4f
Bug fix for dishonest NR method.
