### 2024-07-01 #176 @cmarmo
Hello @jinningwang , I'm Chiara and I'm taking care of the EIC role for the next three months.
I'm just cross-linking here  #174, for future reference.
