### 2024-08-19 ad8e7b4
Follow RTD's deprecation of Sphinx context injection at build time
