### 2024-08-19 b9a1f14
Change itemset to array assignment to ensure compatibility with Numpy 2.0
