+++ andes/variables/dae.py
@@ -472,8 +472,8 @@ class DAE:
         """
         Helper function for setting time in-place.
         """
-
-        self.t = np.array(t, dtype=float)
+        # use in-place assignment
+        self.t[...] = np.array(t, dtype=float)
 
     def get_size(self, name):
         """
