### 2024-08-19 1d14437
Remove redundant comment
