### 2024-09-10 c7e7304
Allow ESD1 to be initialized at negative power by setting lower limit of PHL to `-pmx`.
