### 2024-09-27 ee92612
Fix find_idx to allow multiple matches
