### 2024-09-27 9d3ea43
Add more test on find_idx
