+++ tests/test_model_set.py
@@ -54,3 +54,52 @@ class TestModelMethods(unittest.TestCase):
         ss.GENROU.set("M", np.array(["GENROU_4"]), "v", 6.0)
         np.testing.assert_equal(ss.GENROU.M.v[3], 6.0)
         self.assertEqual(ss.TDS.Teye[omega_addr[3], omega_addr[3]], 6.0)
+
+    def test_find_idx(self):
+        ss = andes.load(andes.get_case('ieee14/ieee14_pvd1.xlsx'))
+        mdl = ss.PVD1
+
+        # multiple values
+        self.assertEqual(mdl.find_idx(keys='name', values=['PVD1_1', 'PVD1_2'],
+                                      allow_none=False, default=False, no_flatten=False),
+                         [1, 2])
+        # multiple values, no flatten
+        self.assertEqual(mdl.find_idx(keys='name', values=['PVD1_1', 'PVD1_2'],
+                                      allow_none=False, default=False, no_flatten=True),
+                         [[1], [2]])
+        # non-existing value
+        self.assertEqual(mdl.find_idx(keys='name', values=['PVD1_999'],
+                                      allow_none=True, default=False, no_flatten=False),
+                         [False])
+
+        # non-existing value is not allowed
+        with self.assertRaises(IndexError):
+            mdl.find_idx(keys='name', values=['PVD1_999'],
+                         allow_none=False, default=False, no_flatten=False)
+
+        # multiple keys
+        self.assertEqual(mdl.find_idx(keys=['gammap', 'name'],
+                                      values=[[0.1, 0.1], ['PVD1_1', 'PVD1_2']]),
+                         [1, 2])
+
+        # multiple keys, with non-existing values
+        self.assertEqual(mdl.find_idx(keys=['gammap', 'name'],
+                                      values=[[0.1, 0.1], ['PVD1_1', 'PVD1_999']],
+                                      allow_none=True, default='CURENT'),
+                         [1, 'CURENT'])
+
+        # multiple keys, with non-existing values not allowed
+        with self.assertRaises(IndexError):
+            mdl.find_idx(keys=['gammap', 'name'],
+                         values=[[0.1, 0.1], ['PVD1_1', 'PVD1_999']],
+                         allow_none=False, default=999)
+
+        # multiple keys, values are not iterable
+        with self.assertRaises(ValueError):
+            mdl.find_idx(keys=['gammap', 'name'],
+                         values=[0.1, 0.1])
+
+        # multiple keys, items length are inconsistent in values
+        with self.assertRaises(ValueError):
+            mdl.find_idx(keys=['gammap', 'name'],
+                         values=[[0.1, 0.1], ['PVD1_1']])
