### 2024-09-27 4d69304
Add test on GroupBase.finx_idx
