### 2024-09-29 eb7d23f
[WIP] Undo changes to pythonapp.yml
