+++ .github/workflows/pythonapp.yml
@@ -8,16 +8,6 @@ jobs:
     runs-on: ubuntu-latest
     steps:
     - uses: actions/checkout@v4
-
-    - name: Log in to Docker Hub
-      run: echo "${{ secrets.DOCKER_PASSWORD }}" | docker login -u "${{ secrets.DOCKER_USERNAME }}" --password-stdin
-
-    - name: Build Docker image
-      run: docker build -t andes:latest .
-
-    - name: Push Docker image
-      run: docker push andes:latest
-
     - uses: conda-incubator/setup-miniconda@v3
       with:
         python-version: 3.11
