### 2024-09-29 e099c9b
[WIP] Fix github action error, add a step to install mamba
