+++ .github/workflows/pythonapp.yml
@@ -17,6 +17,10 @@ jobs:
         channel-priority: true
         activate-environment: anaconda-client-env
     - shell: bash -el {0}
+      name: Install mamba
+      run: |
+        conda install -n base -c conda-forge mamba
+    - shell: bash -el {0}
       name: Install dependencies
       run: |
         mamba install -y nbmake pytest-xdist line_profiler  # add'l packages for notebook tests.
