### 2024-09-29 307bfae
[WIP] Try to fix github action error, use classic solver instead of libmambapy
