+++ .github/workflows/pythonapp.yml
@@ -16,6 +16,8 @@ jobs:
         channels: conda-forge,defaults
         channel-priority: true
         activate-environment: anaconda-client-env
+    - name: Configure conda to use default solver
+      run: conda config --set solver classic
     - shell: bash -el {0}
       name: Install dependencies
       run: |
