### 2024-10-01 e5e19d4
Remove no_flatten in find_idx
