### 2024-10-01 6abf2b4
Fix tests for find_idx
