+++ andes/models/group.py
@@ -264,7 +264,7 @@ class GroupBase:
         Returns
         -------
         list
-            indices of devices
+            A list of lists containing the indices of devices
 
         Examples
         --------
