### 2024-10-01 bccc080
[WIP] Fix find_idx, revert changes
