+++ andes/core/model/modeldata.py
@@ -296,24 +296,7 @@ class ModelData:
         Returns
         -------
         list
-            A list of lists containing the indices of devices
-
-        Examples
-        --------
-        >>> # Use example case of IEEE 14-bus system with PVD1
-        >>> ss = andes.load(andes.get_case('ieee14/ieee14_pvd1.xlsx'))
-
-        >>> # To find the idx of `PVD1` with `name` of 'PVD1_1' and 'PVD1_2'
-        >>> ss.PVD1.find_idx(keys='name', values=['PVD1_1', 'PVD1_2'])
-        [[1], [2]]
-
-        >>> # To find the idx of `PVD1` with `gammap` equals to 0.1
-        >>> ss.PVD1.find_idx(keys='gammap', values=[0.1])
-        ss.PVD1.find_idx(keys='gammap', values=[0.1])
-
-        >>> # To find the idx of `PVD1` with `gammap` equals to 0.1 and `name` of 'PVD1_1'
-        >>> ss.PVD1.find_idx(keys=['gammap', 'name'], values=[[0.1], ['PVD1_1']])
-        [[1]]
+            indices of devices
         """
         if isinstance(keys, str):
             keys = (keys,)
@@ -333,30 +316,21 @@ class ModelData:
             if len(keys) != len(values):
                 raise ValueError("keys and values must have the same length")
 
-            # check if all elements in values have the same length
-            iterator = iter(values)
-            try:
-                first_length = len(next(iterator))
-                ok_len = all(len(element) == first_length for element in iterator)
-            except StopIteration:  # empty iterable
-                ok_len = True
-            if not ok_len:
-                raise ValueError("All elements in values must have the same length")
-
         v_attrs = [self.__dict__[key].v for key in keys]
 
         idxes = []
         for v_search in zip(*values):
-            v_idx_list = []
+            v_idx = None
             for pos, v_attr in enumerate(zip(*v_attrs)):
                 if all([i == j for i, j in zip(v_search, v_attr)]):
-                    v_idx_list.append(self.idx.v[pos])
-            if not v_idx_list:
+                    v_idx = self.idx.v[pos]
+                    break
+            if v_idx is None:
                 if allow_none is False:
                     raise IndexError(f'{list(keys)}={v_search} not found in {self.class_name}')
                 else:
-                    v_idx_list.append(default)
+                    v_idx = default
 
-            idxes.append(v_idx_list)
+            idxes.append(v_idx)
 
         return idxes
