### 2024-10-01 919c8a4
[WIP] Fix find_idx, add parameter allow_all=False to ModelData.find_idx()
