### 2024-10-03 b01fe4d
[WIP] Fix find_idx, add inner lists length check
