### 2024-10-03 e1422b2
[WIP] Fix find_idx, refactor input check
