### 2024-10-03 fdad124
[WIP] Fix find_idx, move input check from modeldata to utils.func
