### 2024-10-03 47e9352
[WIP] Fix find_idx
