### 2024-10-03 ef34488
[WIP] Fix find_idx, minor fix
