### 2024-10-03 33c7ee9
[WIP] Fix find_idx, fix invovled tests
