### 2024-10-03 9a675f5
Update release notes
