### 2024-10-04 3e00282
Add method GroupBase.get_idx() to return all indices in a group
