### 2024-10-04 fbbef5c
Add tests for GroupBase.get_idx()
