### 2024-10-04 6a9d4f2
Update release notes
