### 2024-10-04 a623003
Update release notes
