### 2024-10-05 a9ac7bc
[WIP] ConnMan
