### 2024-10-05 b2265ae
Rename GroupBase.get_dix to GroupBase.get_all_idxes
