+++ tests/test_group.py
@@ -104,7 +104,7 @@ class TestGroup(unittest.TestCase):
         self.assertTrue(any([item == 'y' for item in ff]))
 
         # --- get group idx ---
-        self.assertSetEqual(set(ss.DG.get_idx()),
+        self.assertSetEqual(set(ss.DG.get_all_idxes()),
                             set(ss.PVD1.idx.v))
-        self.assertSetEqual(set(ss.StaticGen.get_idx()),
+        self.assertSetEqual(set(ss.StaticGen.get_all_idxes()),
                             set(ss.PV.idx.v + ss.Slack.idx.v))
