### 2024-10-05 e38166f
Update release notes
