### 2024-10-05 8b4adde
[WIP] ConnMan
