### 2024-10-11 7e330c2
Add a wrapper to check_conn before a routine init
