### 2024-10-11 413ead0
Draft an implementation of ConnMan
