### 2024-10-11 c3475d8
Rename wrapper check_conn as check_conn_before_init
