+++ andes/routines/pflow.py
@@ -6,7 +6,7 @@ import logging
 from collections import OrderedDict
 
 from andes.utils.misc import elapsed
-from andes.routines.base import BaseRoutine, check_conn
+from andes.routines.base import BaseRoutine, check_conn_before_init
 from andes.variables.report import Report
 from andes.shared import np, matrix, sparse, newton_krylov
 
@@ -63,7 +63,7 @@ class PFlow(BaseRoutine):
         self.x_sol = None
         self.y_sol = None
 
-    @check_conn
+    @check_conn_before_init
     def init(self):
         """
         Initialize variables for power flow.
