### 2025-02-12 #176 @lwasser
On hold until https://github.com/pyOpenSci/software-submission/issues/174 is complete.
